Move GenericAdapter to adapters package with name ViewDataBindingAdapter
